source/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto3==1.35.99
botocore==1.35.99
certifi==2024.12.14
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloud-init==24.4
codetiming==1.3.0
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
elastic-transport==8.15.1
elasticsearch==8.17.0
elasticsearch7==7.17.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.3
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-tools==1.45.0
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
MarkupSafe==2.0.1
matplotlib==3.10.0
more-itertools==8.10.0
MouseInfo==0.1.3
netifaces==0.11.0
nose==1.3.7
numpy==2.2.1
oauthlib==3.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pika==1.2.0
pillow==11.1.0
protobuf==3.19.4
psycopg2==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.54
pycurl==7.44.1
pydantic==2.10.4
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.1.1
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==1.0.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-pptx==1.0.2
python-pptx-text-replacer==0.0.6
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
roman==3.3
s3transfer==0.10.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.17.0
sniffio==1.3.1
squarify==0.4.4
ssh-import-id==5.11
starlette==0.41.3
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
uvicorn==0.34.0
wadllib==1.3.6
XlsxWriter==3.2.0
zipp==1.0.0
zope.interface==5.4.0
